extract video player initialization
